Fix startup script to create GS bucket
Added a `gsutil mb` step to the script. The `gsutil cp` command does not automatically create a bucket if it doesn't exist.

Unless the user has a bucket already named with their project ID, the image is never uploaded and the user is never notified that there is an issue.